MVC 패턴으로 코드 변경
